Consolidate pypanda files
Gets rid of the mixins so our generated documentation is cleaner
Change helpers to be panda.helper.arch instead of panda.arch.helper
