--- panda/python/core/panda/arm/__init__.py
@@ -1,5 +1,0 @@
-"""
-Helper functions and information sepecific to the ARM architecture.
-
-It provides things like the register numbers in a convenient way as well as functions specific to ARM.
-"""
