+++ panda/python/core/panda/helper/__init__.py
@@ -1,0 +1,5 @@
+"""
+Helper functions and information sepecific to various architecture.
+
+For supported architectures, the helpers provide the register name->number mappings and a way to dump state
+"""
