+++ panda/python/core/panda/helper/mips.py
