+++ panda/python/core/panda/helper/x86.py
