--- panda/python/core/panda/x86/__init__.py
@@ -1,5 +1,0 @@
-"""
-Helper functions and information sepecific to x86 and x86_64.
-
-It provides things like the register numbers in a convenient way as well as functions specific to x86 and x86_64.
-"""
