+++ panda/python/examples/file_taint/file_taint.py
@@ -14,7 +14,7 @@ import capstone
 from elftools.elf.elffile import ELFFile
 from elftools.elf.sections import SymbolTableSection
 from panda import Panda, blocking, ffi
-from panda.x86.helper import *
+from panda.helper.x86 import *
 
 arch = "x86_64" if len(argv) <= 1 else argv[1]
 extra = "-nographic -chardev socket,id=monitor,path=./monitor.sock,server,nowait -monitor chardev:monitor -serial telnet:127.0.0.1:4445,server,nowait  -device e1000,netdev=net0 -netdev user,id=net0,hostfwd=tcp::5556-:22 -cdrom /home/luke/workspace/qcows/instance-1-cidata.iso"
